@@ -1,32 +1,32 @@
 {
   "name": "next-monaco-example",
-  "version": "0.1.0",
+  "version": "0.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
+    "dev": "next",
     "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "start": "next start"
   },
   "dependencies": {
-    "@graphiql/toolkit": "^0.8.0",
-    "@types/node": "18.13.0",
-    "@types/react": "18.0.28",
-    "@types/react-dom": "18.0.11",
-    "eslint": "8.34.0",
-    "eslint-config-next": "13.1.6",
-    "graphql": "^16.6.0",
-    "graphql-ws": "^5.11.3",
+    "@graphiql/toolkit": "^0.8.4",
+    "graphql": "^16.7.1",
+    "graphql-ws": "^5.14.0",
     "jsonc-parser": "^3.2.0",
     "marked": "^4.2.12",
-    "monaco-editor": "^0.35.0",
+    "monaco-editor": "^0.39.0",
     "monaco-editor-webpack-plugin": "^7.0.1",
-    "monaco-graphql": "^1.1.7",
-    "next": "13.1.6",
-    "next-global-css": "^1.3.1",
-    "prettier": "^2.8.4",
+    "monaco-graphql": "^1.2.4",
+    "next": "13.4.7",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "typescript": "4.9.5"
+    "react-dom": "18.2.0"
+  },
+  "devDependencies": {
+    "eslint": "8.34.0",
+    "@types/node": "18.16.18",
+    "@types/react": "18.2.14",
+    "next-global-css": "1.3.1",
+    "prettier": "3.0.0-alpha.12",
+    "typescript": "5.1.3"
   }
 }
