@@ -1,9 +1,11 @@
-const MonacoWebpackPlugin = require('monaco-editor-webpack-plugin');
-const { patchWebpackConfig } = require('next-global-css');
+import MonacoWebpackPlugin from 'monaco-editor-webpack-plugin';
+import { patchWebpackConfig } from 'next-global-css';
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  reactStrictMode: true,
+  eslint: {
+    ignoreDuringBuilds: true,
+  },
   trailingSlash: true,
   webpack(config, options) {
     // this fixes some issues with loading web workers
@@ -13,7 +15,6 @@ const nextConfig = {
     patchWebpackConfig(config, options);
     config.resolve.alias = {
       ...config.resolve.alias,
-
       // this solves a bug with more recent `monaco-editor` versions in next.js,
       // where vscode contains a version of `marked` with modules pre-transpiled, which seems to break the build.
       //
@@ -36,7 +37,7 @@ const nextConfig = {
               label: 'graphql',
               worker: {
                 id: 'graphql',
-                entry: require.resolve('monaco-graphql/esm/graphql.worker.js'),
+                entry: 'monaco-graphql/esm/graphql.worker.js',
               },
             },
           ],
@@ -49,4 +50,4 @@ const nextConfig = {
   },
 };
 
-module.exports = nextConfig;
+export default nextConfig;
