@@ -1,5 +1,5 @@
-import '../styles/globals.css';
 import type { AppProps } from 'next/app';
+import '../style.css';
 
 export default function App({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />;
